bot/requirements.txt
python-telegram-bot[scheduling]==22.5
python-dotenv==1.2.1
psutil==7.1.3